Add in three setting configurations
This modifies `lib/index.js` to propose three configurations: "loose",
"recommended" and "strict". Recommended was the model for the other two;
loose turns everything down to warning or off levels, while everything
is an error in strict mode.

Of course, I'd really rather recommend basing it on MS style guides, as
I did not do so.

Signed-off-by: Jonathan <jayands.dev@gmail.com>